requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.3.1

# File parsing
pdfplumber==0.11.0
openpyxl==3.1.4
python-docx==1.1.2
pandas==2.2.2

# Embeddings
sentence-transformers==3.0.1
torch==2.3.1
numpy==1.26.4

# Vector store
# hnswlib==0.8.0

# LLM clients
groq==0.9.0
google-generativeai==0.7.2

# Utilities
httpx==0.27.0
loguru==0.7.2
python-dotenv==1.0.1
aiofiles==23.2.1
slowapi==0.1.9

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
